pyproject.toml
[project]
name            = "treineinsite"
version         = "0.1.0"
description     = "Treineinsite - Email sender (Postgres-first)"
readme          = "README.md"
requires-python = ">=3.12"

# (deixa vazio porque vamos instalar tudo via grupos)
dependencies = [
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "typer>=0.15.3",
    "rich>=13.0.0",
    "psycopg[binary]>=3.2",
    "questionary>=2.0.1",
    "prompt_toolkit>=3.0.43",
    "pyfiglet>=0.8.post1",
    "requests>=2.32.0",
    "langchain>=0.2.10",
    "langchain-google-genai>=1.0.5",
    "Babel>=2.15.0",
]

# Entry points de CLI
[project.scripts]
treineinsite-sendemails = "email_sender.cli:main"

# ————————————————————————————
# Grupos de dependência (PEP 735)
# ————————————————————————————
[project.optional-dependencies]
default = [
    "uvicorn[standard]",
    "fastapi",
    "pyyaml>=6.0.2",
    "typer>=0.15.3",
    "rich>=13.0.0",
    "python-dotenv>=1.1.0",
    "flask>=3.0",
    "gunicorn>=22.0",
    "prompt_toolkit>=3.0.43",
    "pyfiglet>=0.8.post1",
]

dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "black>=24.3",
    "isort>=5.13",
    "mypy>=1.10",
    "uvicorn[standard]",  # Add uvicorn to dev dependencies for testing REST API
    "fastapi",  # Add fastapi to dev dependencies for testing REST API
]

# ————————————————————————————
# Build backend
# ————————————————————————————
[build-system]
requires      = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

# ————————————————————————————
# Configs de tooling
# ————————————————————————————
[tool.setuptools]
packages = ["email_sender"]
package-dir = {email_sender = "src/email_sender"}

[tool.black]
line-length    = 88
target-version = ["py312"]
include        = '\.pyi?$'

[tool.isort]
profile            = "black"
multi_line_output  = 3

[tool.mypy]
python_version         = "3.12"
warn_return_any        = true
warn_unused_configs    = true
disallow_untyped_defs  = true

[tool.pytest.ini_options]
addopts = "-q --cov=src/email_sender --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/email_sender"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.coverage.xml]
output = "reports/coverage.xml"

[tool.coverage.html]
directory = "reports/coverage_html"
